requirements.txt
langchain 
pypdf 
pinecone-client 
google-api-python-client 
google-auth 
assemblyai 
elevenlabs
transformers 
torch